test(@lwc/module-resolver): more tests
