+++ packages/@lwc/rollup-plugin/src/__tests__/environment/fixtures/index.js
@@ -1,0 +1,1 @@
+import "lwc";
