+++ packages/@lwc/rollup-plugin/src/__tests__/rootDir/rootDir.spec.js
@@ -1,0 +1,54 @@
+import path from 'path';
+import { rollup } from 'rollup';
+
+import lwc from '../../index';
+
+describe('rootDir', () => {
+    it('warns if an "input" array is passed and when "rootDir" is not set', async () => {
+        const warnings = [];
+
+        await rollup({
+            input: [
+                path.resolve(__dirname, 'fixtures/entryA.js'),
+                path.resolve(__dirname, 'fixtures/entryB.js'),
+            ],
+            plugins: [lwc()],
+            onwarn(warning) {
+                warnings.push(warning);
+            },
+        });
+
+        expect(warnings).toHaveLength(1);
+        expect(warnings[0]).toMatchObject({
+            message: expect.stringMatching(
+                /^The "rootDir" option should be explicitly set when passing an "input" array to rollup\. The "rootDir" option is implicitly resolved to .*\/fixtures.$/
+            ),
+            code: 'PLUGIN_WARNING',
+            plugin: 'lwc',
+        });
+    });
+
+    it('warns if an "input" object is passed and when "rootDir" is not set', async () => {
+        const warnings = [];
+
+        await rollup({
+            input: {
+                entryA: path.resolve(__dirname, 'fixtures/entryA.js'),
+                entryB: path.resolve(__dirname, 'fixtures/entryB.js'),
+            },
+            plugins: [lwc()],
+            onwarn(warning) {
+                warnings.push(warning);
+            },
+        });
+
+        expect(warnings).toHaveLength(1);
+        expect(warnings[0]).toMatchObject({
+            message: expect.stringMatching(
+                /^The "rootDir" option should be explicitly set when passing "input" object to rollup\. The "rootDir" option is implicitly resolved to .*\/fixtures.$/
+            ),
+            code: 'PLUGIN_WARNING',
+            plugin: 'lwc',
+        });
+    });
+});
