+++ packages/@lwc/rollup-plugin/src/__tests__/stylesheetConfig/fixtures/test/test.css
@@ -1,0 +1,3 @@
+:host {
+    color: var(--my-color);
+}
