+++ packages/@lwc/rollup-plugin/src/__tests__/stylesheetConfig/fixtures/test/test.html
@@ -1,0 +1,1 @@
+<template></template>
