+++ packages/@lwc/rollup-plugin/src/__tests__/stylesheetConfig/fixtures/test/test.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class Test extends LightningElement {}
